fix similar bug on functions & revert on loops
It turns out that requiring the loop invariant to be bigger than
the postcondition of the loop is more flexible and allows to get
a tight bound for t62 (in the paper).  This is a bit counter-
intuitive, but it allows some coefficients to be negative.  Also,
loops often exit very close the the invariant.

Making the same modification for function calls does not help and
even makes some simple function tests (t31) fail.